+++ src/ui/ModalFactory.ts
@@ -81,7 +81,7 @@ class TextPromptModal extends Modal {
       this.close()
     }
 
-    new ButtonComponent(btnRow).setButtonText('OK').setCta().onClick(submit)
+    new ButtonComponent(btnRow).setButtonText(t('dialog.ok')).setCta().onClick(submit)
 
     input.addEventListener('keydown', (e) => {
       if (e.key === 'Enter') {